Made the table sortable, and only show the latest search's results
Sorting uses a library: http://www.terrill.ca/sorting/

If you type, for example, "tofu" into the search box, it searches
"t", then "to", then "tof", then "tofu". There are a lot
of foods with "t" in their name, so that result is slow.
There are fewer with "tofu", so that request is fast.
If the "t" request arrives later than the "tofu"
request, then "t" results would still overwrite the old
results.

It's fixed by checking that we're still the most recent
query before filling our results into the table.